v2

+++ graphql/resolvers/bookings.js
@@ -5,8 +5,9 @@ const {
 } = require("../../mongoose/actions/bookings");
 
 const bookingResolver = {
-	bookings: async () => {
+	bookings: async (_, req) => {
 		try {
+			if (!req.isAuth) throw new Error("Unauthorized request");
 			const bookings = await getBookings();
 			return bookings;
 		} catch (err) {
@@ -15,8 +16,9 @@ const bookingResolver = {
 		}
 	},
 
-	bookEvent: async (args) => {
+	bookEvent: async (args, req) => {
 		try {
+			if (!req.isAuth) throw new Error("Unauthorized request");
 			return await createBooking(args.bookingInput);
 		} catch (err) {
 			console.log(err);
@@ -24,8 +26,9 @@ const bookingResolver = {
 		}
 	},
 
-	cancelBooking: async (args) => {
+	cancelBooking: async (args, req) => {
 		try {
+			if (!req.isAuth) throw new Error("Unauthorized request");
 			return await deleteBooking(args.bookingId);
 		} catch (err) {
 			console.log(err);
