+++ graphql/resolvers/events.js
@@ -10,16 +10,18 @@ const eventResolver = {
 			throw err;
 		}
 	},
-	createEvent: async (args) => {
+	createEvent: async (args, req) => {
 		try {
+			if(!req.isAuth) throw new Error("Unauthorized request")
 			return await createEvent(args.eventInput);
 		} catch (err) {
 			console.log(err);
 			throw err;
 		}
 	},
-    deleteEvent: async (args) => {
+    deleteEvent: async (args, req) => {
 		try {
+			if(!req.isAuth) throw new Error("Unauthorized request")
 			return await deleteEvent(args.eventInput);
 		} catch (err) {
 			console.log(err);
